Add files via upload
I was accepted to the Python introductory bootcamp prepared by the Global AI Hub community under the sponsorship of Aygaz. I have successfully completed the Python introductory training that must be completed during the bootcamp and you can find my certificate in the attachment. The bootcamp will continue until August 31, 2024. I will share the project that I will prepare in the bootcamp here at the end of the bootcamp.