Add files via upload
This code simulates a guessing game between the user and a random number by following the steps given. When the user guesses correctly, it ends the game and asks if they want to play again. In this code, while loop, if elif else conditions and basic python coding were used.